backend/requirements.txt
langchain==0.2.16
langchain-groq==0.1.9
fastapi==0.112.0
uvicorn==0.30.5
python-multipart==0.0.9
requests
arxiv
streamlit
python-dotenv
tenacity
reportlab
